[푸시] type 변경 int -> long

+++ app/src/main/java/com/ssomcompany/ssomclient/push/SsomNotiReceiveHandler.java
@@ -34,7 +34,7 @@ public class SsomNotiReceiveHandler implements OneSignal.NotificationReceivedHan
                 Intent localIntent = new Intent(MessageManager.BROADCAST_MESSAGE_RECEIVED_PUSH);
                 localIntent.putExtra(CommonConst.Intent.FROM_USER_ID, data.optString(CommonConst.Intent.FROM_USER_ID, null));
                 localIntent.putExtra(CommonConst.Intent.TO_USER_ID, data.optString(CommonConst.Intent.TO_USER_ID, null));
-                localIntent.putExtra(CommonConst.Intent.TIMESTAMP, data.optInt(CommonConst.Intent.TIMESTAMP, 0));
+                localIntent.putExtra(CommonConst.Intent.TIMESTAMP, data.optLong(CommonConst.Intent.TIMESTAMP, 0));
                 localIntent.putExtra(CommonConst.Intent.CHAT_ROOM_ID, data.optString(CommonConst.Intent.CHAT_ROOM_ID, null));
                 localIntent.putExtra(CommonConst.Intent.MESSAGE, message);
